+++ packages/jest-config/src/index.js
@@ -143,6 +143,7 @@ const getConfigs = (
       skipNodeResolution: options.skipNodeResolution,
       snapshotSerializers: options.snapshotSerializers,
       testEnvironment: options.testEnvironment,
+      testLocationInResults: options.testLocationInResults,
       testMatch: options.testMatch,
       testPathIgnorePatterns: options.testPathIgnorePatterns,
       testRegex: options.testRegex,